differ of iostream and stdio.h

+++ codingTest/2420.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main(){
-    int n,m;
+    long long n,m;
     cin >>n>>m;
     if(n>m) cout<<n-m;
     else cout<<m-n;
